penguin_classification/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
xgboost==2.1.1
google-cloud-storage==2.18.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
locust==2.31.6
numpy==2.1.2
pandas
joblib
scikit-learn